requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
google-cloud-bigquery
google-cloud-storage
python-dotenv